+++ src/index.js
@@ -1,4 +1,3 @@
-
 import {app} from './app.js'
 import dotenv from 'dotenv'
 dotenv.config()
