+++ src/middlewares/multer.js
@@ -1,0 +1,16 @@
+import multer from 'multer'
+import path from 'path'
+
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, path.join('src/public/img'))
+  },
+  filename: (req, file, cb) => {
+    const ext = file.mimetype.split('/')[1]
+    cb(null, file.fieldname + '-' + Date.now() + '.' + ext)
+  }
+})
+
+const upload = multer({storage})
+
+export default upload
