+++ src/routes/category.js
@@ -11,6 +11,7 @@ import {
 
 import {isLoggedIn} from '../libs/logged.js'
 import {verifyToken} from '../middlewares/authjwt.js'
+import upload from '../middlewares/multer.js'
 const router = Router()
 
 // GET categories / private
@@ -22,13 +23,13 @@ router.get('/all/:id', verifyToken, getCategorybyId)
 router.get('/view', isLoggedIn, viewCategory)
 
 // POST CREATE / PRIVATE
-router.post('/add', isLoggedIn, postCategory)
+router.post('/add', isLoggedIn, upload.single('image'), postCategory)
 
 // GET UPADATE / PRIVATE
 router.get('/update/:id', isLoggedIn, updateCategorybyId)
 
 // POST UPADATE / PRIVATE
-router.post('/update/:id', isLoggedIn, postCategorybyId)
+router.post('/update/:id', isLoggedIn, upload.single('image'), postCategorybyId)
 
 // DELETE / PRIVATE
 router.get('/delete/:id', isLoggedIn, deleteCategory)
