+++ src/views/projects/update-project.hbs
@@ -35,6 +35,10 @@
                     <h4>Imágenes</h4>
                     <img class="img-view" src="{{img}}" />
                     <input type="text" name="img" id="blog-img" placeholder="Añadir link" value="{{img}}">
+                    <label class="custom-file-upload btn btn-dark btn-form">
+                        <input type="file" name="image" id="blog-img">
+                        Update File
+                    </label>
                 </div>
             </div>
             <div class="card card-form">
